chatbot_backend/requirements.txt
crewai
crewai-tools
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
langchain-google-genai
cohere


chromadb
fastapi
starlette
uvicorn

pdfplumber
pypdf
pdfminer.six
pypdfium2
beautifulsoup4
pytube

httpx
aiohttp
httpcore
httptools
httpx-sse


pydantic

asyncio
python-dotenv

oauthlib
requests